+++ src/main/kotlin/com/groom/auth/domain/oauth2/CreateOAuth2UserInformation.kt
@@ -1,4 +1,4 @@
-package com.groom.domain.auth
+package com.groom.auth.domain.oauth2
 
 
 data class CreateOAuth2UserInformation(
