+++ src/main/kotlin/com/groom/auth/infrastructure/AuthenticationEntity.kt
@@ -1,7 +1,7 @@
-package com.groom.infrastructure.auth.authentication
+package com.groom.auth.infrastructure
 
-import com.groom.domain.auth.Authentication
-import com.groom.domain.auth.AuthenticationRole
+import com.groom.auth.domain.authentication.Authentication
+import com.groom.auth.domain.authentication.AuthenticationRole
 import com.groom.infrastructure.common.EntityTimeStamp
 import jakarta.persistence.CascadeType
 import jakarta.persistence.Entity
