+++ src/main/kotlin/com/groom/auth/infrastructure/oauth2/OAuth2UserInformationEntity.kt
@@ -1,8 +1,8 @@
-package com.groom.infrastructure.auth.oauth2
+package com.groom.auth.infrastructure.oauth2
 
-import com.groom.domain.auth.CreateOAuth2UserInformation
-import com.groom.domain.auth.OAuth2ProviderName
-import com.groom.domain.auth.OAuth2UserInformation
+import com.groom.auth.domain.oauth2.CreateOAuth2UserInformation
+import com.groom.auth.domain.oauth2.OAuth2ProviderName
+import com.groom.auth.domain.oauth2.OAuth2UserInformation
 import com.groom.infrastructure.common.EntityTimeStamp
 import jakarta.persistence.Entity
 import jakarta.persistence.EnumType
